moved helpers out of test files
